--- tests/unittest/_torch/thop/utils.py
@@ -1,21 +1,0 @@
-import torch
-
-
-def woq_assert_near_eq(ref, act, wTypeId):
-    # match the scale in cpp/tensorrt_llm/kernels/cutlass_kernels/cutlass_preprocessors.cpp
-    if wTypeId == 1:
-        bits_in_type = 8
-    else:
-        bits_in_type = 4
-    quant_range_scale = 1.0 / float(1 << (bits_in_type - 1))
-
-    max_val = torch.max(abs(ref)).item()
-    atol = (max_val * quant_range_scale) * 1.5  # allow for rounding
-    torch.testing.assert_close(ref, act, atol=atol, rtol=1e-7)
-
-
-def woq_groupwise_gt_matmul(mat1, ref_torch_weights, bias=None):
-    ref = torch.matmul(mat1, ref_torch_weights)
-    if bias is not None:
-        ref += bias
-    return ref
